[MT-26] added review fixes on call

+++ MentoringHomeWork/HW11/MotorcycleRepoPattern/MotorcycleDeletedService.cs
@@ -4,6 +4,7 @@ namespace ConsoleApp1;
 public class MotorcycleDeletedService : MotorcycleShowInfo
 {
     private readonly MotorcycleService _motorcycleService = new();
+    
     public void Delete(MotorcycleRepository repository)
     {
         Console.WriteLine("Do you want to delete all data from motorcycle? (y/n)");
@@ -12,21 +13,24 @@ public class MotorcycleDeletedService : MotorcycleShowInfo
 
         if (inputAnswerData.Key == ConsoleKey.Y)
         {
-            var motorcycleToDelete = _motorcycleService.SelectMotorcycle();
+            var motorcycleToDelete = _motorcycleService.SelectMotorcycleById();
+            if (motorcycleToDelete == null)
+            {
+                Console.WriteLine("Motorcycle not found.");
+                Log.Warning("Motorcycle not found.");
+                return;
+            }
+            
             try
             {
                 repository.Delete(motorcycleToDelete);
                 Console.WriteLine($"Motorcycle with ID {motorcycleToDelete.Id} deleted.");
-                Log.Information("motorcycle with ID {id} deleted.",motorcycleToDelete.Id);
+                Log.Information("motorcycle with ID {id} deleted.", motorcycleToDelete.Id);
             }
             catch (Exception e)
             {
                 Log.Error(e.Message);
             }
         }
-
-        Console.WriteLine("Do you want to see All motorcycles? y/n");
-        var input = Console.ReadKey();
-        
     }
 }
